Fuzz the untrusted-input parsers without LLVM, and fix the first findings
tests/managed_integration/fuzz/ has had libFuzzer harnesses for a while, but
they are gated behind -DRETDEC_FUZZ=ON, and fuzz_elf links retdec::fileformat,
which publicly links LLVM. Enabling the option builds the pinned LLVM fork, so
the pull-request job could only check that the harness *files still existed*.
That check passes on a harness that has stopped compiling -- fuzz_dex.cpp had,
calling a DexFile::classDefsSize() that no longer exists.

Nine of the harnesses do not need LLVM at all. The pyc, lua, wasm, dex, apk,
jvm, jar, pdb and cil parsers are in the dependency-free set, so
scripts/standalone_fuzz.sh builds and runs them with clang++ alone. These are
the parsers that read attacker-controlled bytes.

Two modes, deliberately separated. --replay re-runs every seed and every
reproducer under tests/crash_corpus/ with no mutation: deterministic, about a
minute, and it now gates pull requests. Mutation-based discovery runs on a
schedule, writes reproducers into the crash corpus, and uploads them.

Replay passes -malloc_limit_mb, not just -rss_limit_mb. The dominant bug class
here is a parser reserving on a count it read out of the file, which surfaces as
one oversized malloc rather than growing RSS; without the malloc limit those
reproducers replay clean and the entire class goes ungated.

The first four-minute-per-target run found crashes in eight of the nine targets.
Only the CIL parser survived, at 48 million executions. Fixed here:

* pyc: MarshalReader::readTuple reserved on a tuple count taken straight from
  the file -- five header bytes can claim 2^31 elements, 32 GB of shared_ptr.
  readList and readSet had the same unchecked count. All three now reject a
  count larger than the bytes remaining, which no well-formed stream can
  satisfy, and reserve for a small container up front.
* pyc: decodeLnotab311 checked for one byte and read two, and checked for two
  and read three, running off the end of any truncated 3.11 line table.
* lua: eight reserve() calls driven by file-declared counts, plus readInt()
  returning int32_t so a negative count became an enormous size_t, plus
  pos_ + len overflow in three string readers. All now go through
  checkedSize/checkedCount, which measure a claim against the input.

Reproducers for wasm, dex, jvm, jar, apk and pdb are committed ahead of their
fixes so the finds are not lost. fuzz-standalone-replay is therefore red on this
branch until the follow-up commit lands those six.

Also: the standalone check now covers 41 suites over 62 modules -- the parser
suites it had been missing, plus cli_parser, which declares cxx_std_20 in its
own CMakeLists and needed per-module standard support. System zlib is linked for
the suites that read JAR and APK archives. Stray libFuzzer artifacts in the repo
root are ignored; real ones belong in tests/crash_corpus/.

+++ src/pyc_parser/pyc_reader.cpp
@@ -529,17 +529,19 @@ void PycReader::liftInstruction(uint8_t rawOp, int32_t arg,
     case BcOpcode::PYTHON_JUMP_IF_FALSE_OR_POP:
     case BcOpcode::PYTHON_FOR_ITER:
     case BcOpcode::PYTHON_SEND: {
-        // Compute absolute target
-        int32_t target = arg;
+        // Compute absolute target. A chain of EXTENDED_ARG prefixes can carry
+        // any 32-bit operand, so the doubling is done in 64 bits: arg * 2 does
+        // not fit an int32_t for every operand a stream can encode.
+        int64_t target = arg;
         if (info.kind == OpcodeKind::JumpForward)
-            target = static_cast<int32_t>(offset) + 2 + arg * 2;
+            target = static_cast<int64_t>(offset) + 2 + static_cast<int64_t>(arg) * 2;
         else if (info.kind == OpcodeKind::JumpBackward)
-            target = static_cast<int32_t>(offset) + 2 - arg * 2;
+            target = static_cast<int64_t>(offset) + 2 - static_cast<int64_t>(arg) * 2;
         else if (!code.version.atLeast(3, 11))
             target = arg; // absolute offset
         else
-            target = static_cast<int32_t>(offset) + 2 + arg * 2;
-        operand = BcIntOperand{static_cast<int64_t>(target)};
+            target = static_cast<int64_t>(offset) + 2 + static_cast<int64_t>(arg) * 2;
+        operand = BcIntOperand{target};
         break;
     }
     default:
@@ -562,6 +564,11 @@ std::vector<uint32_t> PycReader::findLeaders(const PyCodeObject& code) const {
     const uint8_t haveArg = haveArgument(code.version);
 
     size_t pos = 0;
+    // EXTENDED_ARG prefixes accumulate into an unsigned 32-bit operand, which
+    // is the width the wordcode encodes. Nothing in the stream limits how many
+    // prefixes it chains -- a compiler emits at most three, a hostile file as
+    // many as it likes -- so the accumulator has to wrap at that width instead
+    // of shifting a signed value out of range.
     int32_t extArg = 0;
 
     while (pos < bytecode.size()) {
@@ -591,15 +598,18 @@ std::vector<uint32_t> PycReader::findLeaders(const PyCodeObject& code) const {
 
         OpcodeInfo info = opcodeInfo(op, code.version);
         if (info.isJump() && hasArg) {
-            int32_t target = arg;
+            // Widened for the same reason as in liftInstruction: the operand is
+            // a full 32 bits and pos + arg*2 overruns an int32_t well before it
+            // overruns the bytecode.
+            int64_t target = arg;
             if (info.kind == OpcodeKind::JumpForward)
-                target = static_cast<int32_t>(pos) + static_cast<int32_t>(instrSize)
-                       + arg * 2;
+                target = static_cast<int64_t>(pos) + static_cast<int64_t>(instrSize)
+                       + static_cast<int64_t>(arg) * 2;
             else if (info.kind == OpcodeKind::JumpBackward)
-                target = static_cast<int32_t>(pos) + static_cast<int32_t>(instrSize)
-                       - arg * 2;
+                target = static_cast<int64_t>(pos) + static_cast<int64_t>(instrSize)
+                       - static_cast<int64_t>(arg) * 2;
 
-            if (target >= 0 && static_cast<size_t>(target) < bytecode.size())
+            if (target >= 0 && static_cast<uint64_t>(target) < bytecode.size())
                 leaders.insert(static_cast<uint32_t>(target));
             // Fall-through is a leader too (conditional jumps)
             if (info.kind != OpcodeKind::JumpAbsolute &&
